+++ libft/srcs/ft_abs.c
@@ -6,13 +6,13 @@
 /*   By: rreedy <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/05 20:37:43 by rreedy            #+#    #+#             */
-/*   Updated: 2019/01/05 20:38:22 by rreedy           ###   ########.fr       */
+/*   Updated: 2019/02/28 17:16:40 by rreedy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long int	ft_abs(long int n)
+long long int	ft_abs(long long int n)
 {
 	return ((n < 0) ? n * -1 : n);
 }
